% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_biopax_conversion_chuncks.R
\name{assemble_biopax_conversion_chuncks}
\alias{assemble_biopax_conversion_chuncks}
\title{Assemble Files from BioPAX Conversion Process}
\usage{
assemble_biopax_conversion_chuncks(bindex, wkdir = NULL, output_dir = NULL)
}
\arguments{
\item{bindex}{Number of the source BioPAX object: \cr
1-"BioCarta"\cr
2-"KEGG"\cr
3-"NCI-Nature"\cr
4-"NetPath"\cr
5-"Wiki Pathways"\cr
6-"Science Signaling"\cr
7-"Reactome"\cr
8-"RMC".}

\item{wkdir}{Folder with all files.}

\item{output_dir}{Where to deposit results.}
}
\description{
Assembles all files from BioPAX flattening and re-conversion back to BioPAX, 
when it has been carried out a few pathways at a time, resulting in multiple files per source.
}
\author{
Ivan Grishagin
}
