% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopax_dt_id_cleanup.R
\name{biopax_dt_id_cleanup}
\alias{biopax_dt_id_cleanup}
\title{Clean BioPAX Data Table}
\usage{
biopax_dt_id_cleanup(biopax_dt)
}
\arguments{
\item{biopax_dt}{BioPAX-style datatable.}
}
\description{
Takes in a BioPAX-style datatable, 
and cleans it by removing duplicate entities, that have the same values, but different ids.
}
\details{
Outputs a BioPAX-style datatable.
}
\author{
Ivan Grishagin
}
