% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_biopax.R
\name{check_biopax}
\alias{check_biopax}
\title{Check BioPAX by Plotting Certain Pathways}
\usage{
check_biopax(biopax, pw_to_check = NULL, what = c("sample", "order", "id"),
  output_identifier = NULL, verbose = TRUE)
}
\arguments{
\item{biopax}{BioPAX object.}

\item{pw_to_check}{Which pathways to check? Can take in (1) pathway ids or
(2) integer as number of pathways to randomly select and check 
(3) vector of integers as pathway indices.}

\item{what}{What's the format of supplied ids? Either \code{id}, or \code{sample}, or \code{order}.}

\item{output_identifier}{What to append to output file?}

\item{verbose}{Logical. Show all pertaining progress?}
}
\description{
Plot regulatory graph for each supplied pathway.
}
\author{
Ivan Grishagin
}

