% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_biopax_interactions_vect.R
\name{extract_biopax_interactions_vect}
\alias{extract_biopax_interactions_vect}
\title{Extract BioPAX Interactions}
\usage{
extract_biopax_interactions_vect(pw_id_vect = "all", biopax, biopax_source,
  tag = NULL)
}
\arguments{
\item{pw_id_vect}{Vector of pathway ids to process. Can also take a value \code{all} to process all pathways.}

\item{biopax}{BioPAX object.}

\item{biopax_source}{Name of the public source of the BioPAX object.}

\item{tag}{String to append to the output file.}
}
\description{
Takes in a BioPAX object and returns a dataframe of all "interactions", one per row, for a specified vector of pathways.
}
\details{
Also outputs the resultant dataframe as *.txt and *.xlsx files.
}
\author{
Ivan Grishagin
}
