% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBiopaxInteractions.R
\name{extractBiopaxInteractions}
\alias{extractBiopaxInteractions}
\title{Extract Pathway Interactions}
\usage{
extractBiopaxInteractions(pw_id, biopax, biopax_source)
}
\arguments{
\item{pw_id}{Pathway id to process.}

\item{biopax}{BioPAX object.}

\item{biopax_source}{Name of the public source of the BioPAX object.}
}
\description{
Takes in a BioPAX object and returns a dataframe of all "interactions", one per row, for ONE specified pathway.
}
\author{
Ivan Grishagin
}
