% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions_df_to_biopax_dt.R
\name{interactions_df_to_biopax_dt}
\alias{interactions_df_to_biopax_dt}
\title{Convert Interactions Dataframe into BioPAX Data Table}
\usage{
interactions_df_to_biopax_dt(interactions_df)
}
\arguments{
\item{interactions_df}{\code{MAIN_flatten_biopax} output dataframe.}
}
\description{
Convert dataframe of interactions into a BioPAX-style data table.
}
\details{
Takes in a dataframe returned by \code{MAIN_flatten_biopax}, 
and parses each value to return a BioPAX-style data table.
}
\author{
Ivan Grishagin
}

