% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_biopax_comparison.R
\name{MAIN_biopax_comparison}
\alias{MAIN_biopax_comparison}
\title{MAIN -- Compare BioPAX Objects}
\usage{
MAIN_biopax_comparison(pwid_to_compare = c("all", "none"),
  pwid_to_plot = c("none", "same", "test5"), new_biopax, orig_biopax,
  new_biopax_tag = "_new", orig_biopax_tag = "_orig", verbose = TRUE,
  pw_df_path = "default", source_name = NULL)
}
\arguments{
\item{pwid_to_compare}{Which pathway IDs to use for the comparison of control components? Takes either a vector of pathway IDs, 
or \code{all} to process all pathways, or \code{none} to not use this comparison at all.}

\item{pwid_to_plot}{Which pathways to plot for subsequent visual inspection? Takes either a vector of pathway IDs, 
or \code{same} to process the same pathways as for control component comparison, 
or \code{test5} to plot 5 random pathways (can change to any number),
or \code{none} to not use this comparison at all.}

\item{new_biopax}{First BioPAX object.}

\item{orig_biopax}{Second BioPAX object.}

\item{new_biopax_tag}{First BioPAX Object tag.}

\item{orig_biopax_tag}{BioPAX object.}

\item{verbose}{Logical. Show all pertaining progress?}

\item{pw_df_path}{}

\item{source_name}{BioPAX source name.}
}
\description{
Compare pathways in two BioPAX objects.
}
\details{
Allows to compare control components
}
\author{
Ivan Grishagin
}

