% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_combine_clean_biopax_unifyids.R
\name{MAIN_combine_clean_biopax_unifyids}
\alias{MAIN_combine_clean_biopax_unifyids}
\title{MAIN -- Combine and Clean List of BioPAX Objects}
\usage{
MAIN_combine_clean_biopax_unifyids(new_biopax_list, pw_df_path = "default",
  exclude_id_pattern = "inxight_pathways")
}
\arguments{
\item{new_biopax_list}{List of BioPAX objects.}

\item{pw_df_path}{Path to the source table with pathway, id, name, and source values.}

\item{exclude_id_pattern}{Exclude pathways with such pattern in their IDs from ID unification.}
}
\description{
Merges BioPAX objects, unifies component IDs, replaces certain pathway ids with proper \code{inxight_pathway} ones, 
and cleans BioPAX objects from html tags and garbled symbols.
}
\details{
Wrapper for \code{unify_biopax_ids} and \code{biopax_from_dt}.
ID unification is conducted by appending a class of a component to its throughout number.
E.g. for a component of class Protein, #11 from the top of the table, ID will be "Protein11".
}
\author{
Ivan Grishagin
}
