% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_flatten_biopax.R
\name{MAIN_flatten_biopax}
\alias{MAIN_flatten_biopax}
\title{MAIN -- Flatten BioPAX}
\usage{
MAIN_flatten_biopax(pwid_to_convert = c("test5", "all"),
  work_dir = "default", pw_df_path = "default", biopax = NULL,
  source_name = NULL, groupnum = NULL, ngroups = NULL)
}
\arguments{
\item{pwid_to_convert}{Which pathway IDs to process? Takes either a vector of pathway IDs, 
or \code{test5} to process 5 random pathways (can change to any number), 
or \code{all} to process all pathways.}

\item{pw_df_path}{Path to the source table with pathway, id, name, and source values.}

\item{biopax}{BioPAX object.}

\item{source_name}{BioPAX public source name.}

\item{groupnum}{Split pathway ids into \code{ngroups} groups. 
Only process group #\code{groupnum} out of a total of \code{ngroups}.}

\item{ngroups}{See \code{groupnum}.}
}
\description{
Flatten BioPAX ontology file by extracting all "interactions".
}
\details{
Returns a dataframe with one complete "interaction" (pathway component) per row. 
Each "interaction" contains over 120 columns, flattening the BioPAX ontology as much as it is practically feasible. 
Each complex component is represented as a special string structure, similar to JSON.
}
\author{
Ivan Grishagin
}
