% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_write_biopax.R
\name{MAIN_write_biopax}
\alias{MAIN_write_biopax}
\title{MAIN -- Write BioPAX}
\usage{
MAIN_write_biopax(combined_biopax, what = c("both", "onebig", "manysmall"),
  onebigfilename = NULL, pwtoextract_pattern = "inxight_pathways",
  pwtoextract = "default", bigtarname = "default",
  smalltarname = "default")
}
\arguments{
\item{combined_biopax}{BioPAX object.}

\item{what}{How to write: \code{onebig} (all pathways in one file), \code{manysmall} (one pathway per file) or \code{both}.}

\item{onebigfilename}{In case one file with all pathways desired, which name should be used for it?}

\item{pwtoextract_pattern}{Only pathways with IDs containing the pattern will be extracted.}

\item{pwtoextract}{IDs of pathways to extract. Trumps \code{pwtoextract_pattern}.}

\item{bigtarname}{Name of the tar.gz archive with one large pathway file.}

\item{smalltarname}{Name of the tar.gz archive with individual pathway files.}
}
\description{
Wrapper for a \code{biopax_from_dt} or \code{splitbiopax_from_dt} functions.
}
\details{
Returns either a graph or a dataframe with control-type components.
}
\author{
Ivan Grishagin
}
