% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway2RegulatoryGraph_Rancho.R
\name{pathway2RegulatoryGraph_Rancho}
\alias{pathway2RegulatoryGraph_Rancho}
\title{Pathway to Regulatory Graph}
\usage{
pathway2RegulatoryGraph_Rancho(biopax, pwid, expandSubpathways = TRUE,
  splitComplexMolecules = TRUE, useIDasNodenames = FALSE, verbose = FALSE,
  returnGraph = TRUE)
}
\arguments{
\item{biopax}{BioPAX object.}

\item{pwid}{Pathway ID.}

\item{expandSubpathways}{Logical. Expand sub-pathways? Defaults to \code{TRUE}.}

\item{splitComplexMolecules}{Logical. Split complex molecules into individual components? Defaults to \code{TRUE}.}

\item{useIDasNodenames}{Logical. Use node ids as names? Defaults to \code{FALSE}.}

\item{verbose}{Logical. Show all pertaining progress? Defaults to \code{FALSE}.}

\item{returnGraph}{Logical. Return graph (as opposed to a dataframe of control-type components)? Defaults to \code{TRUE}.}
}
\description{
A different interpretation of a \code{pathway2RegulatoryGraph} function from rBiopaxParser package.
}
\details{
Returns either a graph or a dataframe with control-type components.
}
\author{
Ivan Grishagin
}

