% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RIGconvertbiopax.R
\name{update_RIGconvertbiopax}
\alias{update_RIGconvertbiopax}
\title{Update RIGconvertbiopax from GitHub}
\usage{
update_RIGconvertbiopax()
}
\description{
Updates the package. It may be necessary to restart the R session after the update.
}
\author{
Ivan Grishagin
}

