% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_flatten_biopax.R
\name{MAIN_flatten_biopax}
\alias{MAIN_flatten_biopax}
\title{MAIN -- Flatten BioPAX}
\usage{
MAIN_flatten_biopax(pwid_to_convert = c("test5", "all"),
  work_dir = "D:/Dropbox/Rancho/NCATS/ToxDB/",
  pw_df_path = "./_source_files/pathways_matched_to_sources_v015.xlsx",
  biopax = NULL, source_name = NULL, groupnum = NULL, ngroups = NULL)
}
\arguments{
\item{pwid_to_convert}{Which pathway IDs to process? Takes either a vector of pathway IDs, 
or \code{test5} to process 5 random pathways (can change to any number), 
or \code{all} to process all pathways.}

\item{combined_biopax}{BioPAX object.}

\item{what}{How to write: \code{onebig} (all pathways in one file), \code{manysmall} (one pathway per file) or \code{both}.}

\item{onebigfilename}{In case one file with all pathways desired, which name should be used for it?}

\item{pwtoextract_pattern}{Only pathways with IDs containing the pattern will be extracted.}
}
\description{
Flatten BioPAX ontology file by extracting all "interactions".
}
\details{
Returns a dataframe with one complete "interaction" (pathway component) per row. 
Each "interaction" contains over 120 columns, flattening the BioPAX ontology as much as it is practically feasible. 
Each complex component is represented as a special string structure, similar to JSON.
}
\author{
Ivan Grishagin
}

